pyproject.toml
[tool.poetry]
name = "mazgen2.0"
version = "2.0.0"
description = "Simple website that generates a pen and paper maze"
authors = ["Jasper Wagen <jasperwagen01@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
Flask = "^2.2.2"
Pillow = "^9.2.0"
numpy = "^1.23.2"
waitress = "^2.1.2"

[tool.poetry.dev-dependencies]
isort = "^5.10.1"
black = "^22.6.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
